refactor: Split out CommandResultSummaryView
